% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20021R.R
\name{dqqplot}
\alias{dqqplot}
\title{A qqplot comparing the bootstrap values of the empirical log likelihood of the mean in the original two-dimensional sample to the chi-square distribution with 2 degree using R.}
\usage{
dqqplot(data, B)
}
\arguments{
\item{data}{the original sample in n*2 matrix}

\item{B}{the number of bootstrap times}
}
\value{
a qqplot
}
\description{
The theorem could be found in the book Empirical Likelihood by Art B.Owen.
}
\examples{
\dontrun{
data(data)
attach(data)
dpplot(data,1000)
}
}
