% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20021R.R
\name{f}
\alias{f}
\title{Use four inputs to calculate the empirical log likelihood of the mean of the original two-dimensional sample using R.}
\usage{
f(x, y, xbar, ybar)
}
\arguments{
\item{x}{the bootstrap sample in the first dimension}

\item{y}{the bootstrap sample in the second dimension}

\item{xbar}{the mean of the original sample in the first dimension}

\item{ybar}{the mean of the original sample in the second dimension}
}
\value{
the empirical log likelihood of the mean in the original sample
}
\description{
The theorem could be found in the book Empirical Likelihood by Art B.Owen.
}
