% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolis}
\alias{rwMetropolis}
\title{A  random walk Metropolis sampler using Rcpp}
\usage{
rwMetropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance of the normal distribution}

\item{x0}{the initial value}

\item{N}{the number of points}
}
\value{
a random sample of size \code{n}
}
\description{
A random walk Metropolis sampler using Rcpp
}
