\name{StatComp20021-package}
\alias{StatComp20021-package}
\alias{StatComp20021}
\docType{package}
\title{
  This package verifies a theorem about empirical log likelihood.
}
\description{
  This package caiculate the empirical log likelihood of the mean of the original two-dimensional sample using R,and then make a qqplot for comparison and vertification.
}
\author{
Chi Sun

Maintainer: Chi Sun <sc1997@mail.ustc.edu.cn>
}
\keyword{ package }


